+++ lib/types/EasyNativeBinding.ts
@@ -32,6 +32,7 @@ import { CurlInfoDebug } from '../enum/CurlInfoDebug'
 import { CurlIpResolve } from '../enum/CurlIpResolve'
 import { CurlNetrc } from '../enum/CurlNetrc'
 import { CurlPause } from '../enum/CurlPause'
+import { CurlPreReqFunc } from '../enum/CurlPreReqFunc'
 import { CurlProgressFunc } from '../enum/CurlProgressFunc'
 import { CurlProtocol } from '../enum/CurlProtocol'
 import { CurlProxy } from '../enum/CurlProxy'
@@ -253,6 +254,24 @@ export declare class EasyNativeBinding {
           cacheEntry: CurlHstsCacheEntry,
           cacheCount: CurlHstsCacheCount,
         ) => any)
+      | null,
+  ): CurlCode
+  /**
+   * Use {@link "Curl".Curl.option|`Curl.option`} for predefined constants.
+   *
+   *
+   * Official libcurl documentation: [`curl_easy_setopt()`](http://curl.haxx.se/libcurl/c/curl_easy_setopt.html)
+   */
+  setOpt(
+    option: 'PREREQFUNCTION',
+    value:
+      | ((
+          this: EasyNativeBinding,
+          connPrimaryIp: string,
+          connLocalIp: string,
+          connPrimaryPort: number,
+          conLocalPort: number,
+        ) => CurlPreReqFunc)
       | null,
   ): CurlCode
   /**